integrator/requirements-demonstrator.txt
docopt==0.6.2
openpyxl==3.0.9
python-docx==0.8.11
sortedcontainers==2.4.0

mammoth
ipysheet
